requirements.txt
crewai
langchain_groq
langchain_community.llms
langchain-google-genai